完成运费模版单元测试，并恢复权限中心和会员中心代码

+++ src/main/java/com/zpl/eshop/membership/service/MembershipService.java
@@ -13,9 +13,9 @@ import java.util.List;
 public interface MembershipService {
 
     /**
-     * 通知会员中心，“用户今日第一次登陆”事件发生了
+     * 通知会员中心，“用户今日第一次登录”事件发生了
      *
-     * @param userAccountId 用户帐号 id
+     * @param userAccountId 用户账号ID
      * @return 处理结果
      */
     Boolean informFirstLoginDailyEvent(Long userAccountId);
@@ -23,25 +23,25 @@ public interface MembershipService {
     /**
      * 通知会员中心，“支付订单”事件发生了
      *
-     * @param userAccountId    用户帐号 id
+     * @param userAccountId    用户账号id
      * @param totalOrderAmount 订单总金额
      * @return 处理结果
      */
-    Boolean informPayOrderEvent(Long userAccountId, Double totalOrderAmount);
+    Boolean informPayOrderEvent(Long userAccountId, Long totalOrderAmount);
 
     /**
      * 通知会员中心，“完成退货”事件发生了
      *
-     * @param userAccountId    用户帐号 id
+     * @param userAccountId    用户账号id
      * @param totalOrderAmount 订单总金额
      * @return 处理结果
      */
-    Boolean informFinishReturnGoodsEvent(Long userAccountId, Double totalOrderAmount);
+    Boolean informFinishReturnGoodsEvent(Long userAccountId, Long totalOrderAmount);
 
     /**
      * 通知会员中心，“发表评论”事件发生了
      *
-     * @param userAccountId 用户帐号 id
+     * @param userAccountId 用户账号id
      * @param showPictures  是否晒图
      * @return 处理结果
      */
@@ -50,7 +50,7 @@ public interface MembershipService {
     /**
      * 通知会员中心，“删除评论”事件发生了
      *
-     * @param userAccountId 用户帐号 id
+     * @param userAccountId 用户账号id
      * @param showPictures  是否晒图
      * @return 处理结果
      */
@@ -59,7 +59,8 @@ public interface MembershipService {
     /**
      * 查询所有的用户账户
      *
-     * @return 用户账户
+     * @return
      */
     List<UserAccountDTO> listAllUserAccounts();
+
 }
